rewrote the entire api for speeches to make it so that it suits the new speech logic
